+++ dls_barcode/scan/open/__init__.py
@@ -1,1 +1,1 @@
-from .scan import OpenScanner
+from .open_scanner import OpenScanner
