+++ dls_barcode/scan/with_geometry/__init__.py
@@ -1,2 +1,2 @@
-from .scan_slot import SlotScanner
-from .scan import GeometryScanner
+from .slot_scanner import SlotScanner
+from .geometry_scanner import GeometryScanner
