+++ dls_barcode/scan/with_geometry/slot_scanner.py
